Add support for 204 response (#5)
In the case of a 204 status response, there will be no data frame sent and the headers frame will have the endofstream flag set (assuming there are no continuation frames). This update will check to see if the endofstream flag is set in the headers frame and complete the stream/result if it is set.
Additionally, updated travis ci.